requirements.txt
bitsandbytes==0.42.0
deepspeed==0.15.4
editdistance==0.8.1
einops==0.8.1
fastapi==0.115.12
flash_attn==2.7.3
gradio==5.34.0
huggingface_hub==0.27.1
llava==1.2.2.post1
numpy==2.3.0
openai==1.86.0
opencv_python==4.11.0.86
packaging==25.0
petrel_oss_sdk==v2.2.1_2_g1505ef3_master
Pillow==11.2.1
pycocoevalcap==1.2
ray==2.47.0
Requests==2.32.4
scipy==1.15.3
sentencepiece==0.1.99
sglang==0.4.7
shortuuid==1.0.13
tensorflow==2.19.0
timm==1.0.15
tokenizers==0.21.1
torch==2.1.2
torchvision==0.16.2
tqdm==4.67.1
transformers==4.45.0
uvicorn==0.34.3
xformers==0.0.30
